Moved user pages to their own folder
