+++ frontend/src/UserPages/EditTherapist.js
@@ -1,6 +1,6 @@
 import React from "react";
-import "./FormStyles.css";
-import Nav from "./Nav";
+import "../FormStyles.css";
+import Nav from "../Nav";
 
 class EditTherapist extends React.Component {
